apps/admin-web/7-multi-agent/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.34.0
pydantic>=2.5.0
google-adk==0.1.0
requests>=2.31.0
